Reorganize resstock fetch/copy scripts
