+++ data/resstock/nrel/copy_resstock_data.py
